requirements.txt
Flask
pypandoc
requests
tiktoken
PyYAML
# -e .